added id to parachain and bp charts for substrate
